feat(resize-canvas): add absolute sizing
Add an option to size the canvas with absolute values. This has the
unfortunate symptom of messing with the parameters, so biting the
bullet and changing the api.

BREAKING CHANGE: resize-canvas now accepts a single config/spec object
instead of individualized parameters. Upgrade path is as follows.

Before:
resizeCanvas(canvas, [20, 20], [0, 0])

After:
resizeCanvas({
  canvas: canvas,
  diff: [20, 20], // or now optionally you can use {size} to set an
absoulte new size
  from: [0, 0]
})